+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/infrastructure/di/CoreModule.kt
@@ -11,7 +11,7 @@ import com.maksimowiczm.foodyou.feature.home.mealscard.ui.app.product.update.Upd
 import com.maksimowiczm.foodyou.feature.home.mealscard.ui.app.search.SearchViewModel
 import com.maksimowiczm.foodyou.feature.home.mealscard.ui.card.MealsCardViewModel
 import com.maksimowiczm.foodyou.feature.settings.goalssettings.ui.GoalsSettingsViewModel
-import com.maksimowiczm.foodyou.feature.settings.mealssettings.newui.MealsSettingsScreenViewModel
+import com.maksimowiczm.foodyou.feature.settings.mealssettings.ui.MealsSettingsScreenViewModel
 import com.maksimowiczm.foodyou.ui.DiaryViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.bind
